pesona jember pra release

+++ app/Livewire/Order/CompleteTransaction.php
@@ -3,6 +3,7 @@
 namespace App\Livewire\Order;
 
 use App\Models\Order;
+use App\Models\OrderSharingDetail;
 use Livewire\Component;
 
 class CompleteTransaction extends Component
@@ -10,28 +11,50 @@ class CompleteTransaction extends Component
     public $orderId;
     public $order;
     public $total;
+
 //    public $
-    public function mount(){
+    public function mount()
+    {
         $this->order = Order::find($this->orderId);
-$this->getAll();
+
     }
 
-//    public function getAll(){
-//        foreach ($this->order->orderProducts as $orderProduct){
-//            if ($this->order->transaction_type_id!=3){
-//                $orderProduct->value-$orderProduct->-getTax($orderProduct->value,$this->order->ppn,$orderProduct->pph);
-//            }
-//        }
-//    }
+    public function getAll()
+    {
+
+        if ($this->order->transaction_type_id != 3) {
+
+            $afterTax = 0;
+            foreach ($this->order->orderProducts as $op) {
+                $dpp = $op->value * 100 / (100 + $this->order->ppn);
+                $ppnProduct = $op->value - $dpp;
+                $pphProduct = $op->pph * $dpp / 100;
+                $afterTax += ($dpp - $pphProduct);
+            }
+
+            $allSharing = 0;
+            foreach ($this->order->orderProducts as $item2) {
+                $afterTax += getTax($item2->value, $this->order->ppn, $item2->pph);
+                foreach ($this->order->orderSharings as $s) {
+                    $osd = OrderSharingDetail::where('order_sharing_id', $s->id)->where('order_product_id', $item2->id)->first();
+                    if ($osd != null) {
+                        $allSharing += $osd->percentage * getTax(($item2->price * $item2->quantity), $this->order->ppn, $item2->pph) / 100;
+                    }
+                }
+            }
+            $allHppValue = $this->order->orderProducts->sum('hpp_value');
+            $margin = $afterTax - $allHppValue - $allSharing;
+
+            return $margin;
+        } else {
+
+            return $this->order->value * $this->order->percentage / 100;
+        }
+    }
 
     public function endTransaction()
     {
-        $this->dispatch('swal:confirm', data: [
-            'icon' => 'warning',
-            'title' => 'apakah anda yakin ingin menyelesaikan transaksi ini, setelah anda menyelesaikan anda tidak bisa melakukan perubahan lagi',
-            'confirmText' => 'Selesaikan',
-            'method' => 'delete',
-        ]);
+        $this->dispatch('swal:confirm', data: ['icon' => 'warning', 'title' => 'apakah anda yakin ingin menyelesaikan transaksi ini, setelah anda menyelesaikan anda tidak bisa melakukan perubahan lagi', 'confirmText' => 'Selesaikan', 'method' => 'delete',]);
 
     }
